BAU: Consolidate all existing login tests into one file
We no longer deal with the authentication attempts service in the login handler, so the remaining two tests in this file can be moved into the main login handler test file.

At some point we should consolidate all the tests in this file and group better, but for now introduce a nested Reauth class for these to separate them from the other tests
